Fix git working_dir parent walk when run in repo's top-level dir.
Started an abstract CatalogRef class and 3 subclasses.

This is related to refactoring the current provider
code into two envisioned parts...

A) CatalogRef -- initializing a concrete CatalogRef subclass (whether
for a local catalog or for a DB scope or a composite catalog),
updating that CatalogRef instance with newer data, and
searching/find()'ing ToolDescriptors from a CatalogRef instance...
are envisioned to be part of (or near) this newly introduced
core/catalog/ref.py file.

B) Processing a single ToolDescriptor (which may or may not have come
from the results of a CatalogRef.search/find()) into a callable python
function instance (which might in turn be then passed onwards to a
langchain-or-equivalent framework) is a separate, orthogonal maneuver,
which will likely stay in the core/tool/provider.py file or nearby.

Renamed to MemCatalogRef (to handle dirty case).

Started sketching out cmd_find.py implementation.

More sketching of catalog/refs.py methods.

MemCatalogRef.save() method.