Annotations at find-time are now specified using a query string.
By default, we'll have an empty metadata field here.

Cleaning up the output for the find command (the index_catalog for include_dirty now writes the debug logger instead of the console) and we are taking advantage of the verbose option when necessary.

Update to AnnotationPredicate to properly handle None.

Found a bug with the HTTP request generator on using tools straight from the catalog. We now define a handle that grabs the OpenAPI spec first.

Adding lock file